+++ src/ui_components/AddLayerButton/AddLayerButton.tsx
@@ -3,17 +3,50 @@ import React from 'react';
 import PlaylistAddIcon from '@mui/icons-material/PlaylistAdd';
 import AddLayerPopup from './AddLayerPopup';
 import { LayerConfig } from '../../contexts/LayerContext';
+import useAddProtocol, { useAddProtocolProps } from '../../hooks/useAddProtocol/useAddProtocol';
+import { mbTilesProtocolHandler } from '../../protocol_handlers/mbtiles';
 
+function AddProtocol(props: useAddProtocolProps) {
+	useAddProtocol({
+		protocol: props.protocol,
+		handler: props.handler,
+	});
+	return <></>;
+}
 export interface AddLayerButtonProps {
 	sx?: SxProps;
+	/**
+	 * An string array, to filter the supported file types that would be shown to the user
+	 * Default is: ['geojson', 'wms', 'csv', 'topojson', 'osm', 'gpx', 'kml', 'tcx']
+	 */
+	layerTypes?: string[];
 	onComplete?: (config: LayerConfig) => void;
 }
 
 const AddLayerButton = (props: AddLayerButtonProps) => {
 	const [popupOpen, setPopupOpen] = React.useState<boolean>(false);
+	const layerTypes = props.layerTypes || [
+		'geojson',
+		'wms',
+		'mbtiles',
+		'csv',
+		'topojson',
+		'osm',
+		'gpx',
+		'kml',
+		'tcx',
+	];
 
 	return (
 		<>
+			{layerTypes.includes('mbtiles') && (
+				<AddProtocol
+					{...{
+						protocol: 'mbtiles',
+						handler: mbTilesProtocolHandler,
+					}}
+				/>
+			)}
 			<Button
 				variant="contained"
 				sx={{ marginTop: '10px', ...props.sx }}
@@ -21,7 +54,12 @@ const AddLayerButton = (props: AddLayerButtonProps) => {
 			>
 				<PlaylistAddIcon />
 			</Button>
-			<AddLayerPopup open={popupOpen} setOpen={setPopupOpen} onComplete={props?.onComplete} />
+			<AddLayerPopup
+				open={popupOpen}
+				setOpen={setPopupOpen}
+				onComplete={props?.onComplete}
+				layerTypes={layerTypes}
+			/>
 		</>
 	);
 };